Fix buildind on ubuntu

+++ tests/social_index_calculator.cpp
@@ -62,12 +62,12 @@ BOOST_AUTO_TEST_CASE( test1 )
     auto p_account_index_map = r[node_type::ACCOUNT];
     auto p_content_index_map = r[node_type::CONTENT];
     
-    BOOST_CHECK_CLOSE(p_account_index_map->at("account-0"), 0.1826260, 1e-3);
-    BOOST_CHECK_CLOSE(p_account_index_map->at("account-1"), 0.4055207, 1e-3);
-    BOOST_CHECK_CLOSE(p_account_index_map->at("account-2"), 0.4118532, 1e-3);
-    BOOST_CHECK_CLOSE(p_content_index_map->at("post-0"), 0.2689898, 1e-3);
-    BOOST_CHECK_CLOSE(p_content_index_map->at("post-1"), 0.3560182, 1e-3);
-    BOOST_CHECK_CLOSE(p_content_index_map->at("post-2"), 0.3749918, 1e-3);
+    BOOST_CHECK_CLOSE((double) p_account_index_map->at("account-0"), 0.1826260, 1e-3);
+    BOOST_CHECK_CLOSE((double) p_account_index_map->at("account-1"), 0.4055207, 1e-3);
+    BOOST_CHECK_CLOSE((double) p_account_index_map->at("account-2"), 0.4118532, 1e-3);
+    BOOST_CHECK_CLOSE((double) p_content_index_map->at("post-0"), 0.2689898, 1e-3);
+    BOOST_CHECK_CLOSE((double) p_content_index_map->at("post-1"), 0.3560182, 1e-3);
+    BOOST_CHECK_CLOSE((double) p_content_index_map->at("post-2"), 0.3749918, 1e-3);
     
     double_type account_sum = p_account_index_map->at("account-0") + p_account_index_map->at("account-1") + p_account_index_map->at("account-2");
     
@@ -75,11 +75,11 @@ BOOST_AUTO_TEST_CASE( test1 )
     
     double_type total_sum = content_sum + account_sum;
     
-    BOOST_CHECK_CLOSE(account_sum, 1, 1e-3);
+    BOOST_CHECK_CLOSE((double) account_sum, 1, 1e-3);
 
-    BOOST_CHECK_CLOSE(content_sum, 1, 1e-3);
+    BOOST_CHECK_CLOSE((double) content_sum, 1, 1e-3);
 
-    BOOST_CHECK_CLOSE(total_sum, 2, 1e-3);
+    BOOST_CHECK_CLOSE((double) total_sum, 2, 1e-3);
     
     
     BOOST_CHECK_EQUAL(calculator->get_total_handled_block_count(), 1);
@@ -99,12 +99,12 @@ BOOST_AUTO_TEST_CASE( test2 )
     auto p_account_index_map = r[node_type::ACCOUNT];
     auto p_content_index_map = r[node_type::CONTENT];
     
-    BOOST_CHECK_CLOSE(p_account_index_map->at("account-0"), 0.1834468, 1e-3);
-    BOOST_CHECK_CLOSE(p_account_index_map->at("account-1"), 0.4080653, 1e-3);
-    BOOST_CHECK_CLOSE(p_account_index_map->at("account-2"), 0.4084877, 1e-3);
-    BOOST_CHECK_CLOSE(p_content_index_map->at("post-0"), 0.2704000, 1e-3);
-    BOOST_CHECK_CLOSE(p_content_index_map->at("post-1"), 0.3624122, 1e-3);
-    BOOST_CHECK_CLOSE(p_content_index_map->at("post-2"), 0.3671877, 1e-3);
+    BOOST_CHECK_CLOSE((double) p_account_index_map->at("account-0"), 0.1834468, 1e-3);
+    BOOST_CHECK_CLOSE((double) p_account_index_map->at("account-1"), 0.4080653, 1e-3);
+    BOOST_CHECK_CLOSE((double) p_account_index_map->at("account-2"), 0.4084877, 1e-3);
+    BOOST_CHECK_CLOSE((double) p_content_index_map->at("post-0"), 0.2704000, 1e-3);
+    BOOST_CHECK_CLOSE((double) p_content_index_map->at("post-1"), 0.3624122, 1e-3);
+    BOOST_CHECK_CLOSE((double) p_content_index_map->at("post-2"), 0.3671877, 1e-3);
     
     double_type account_sum = p_account_index_map->at("account-0") + p_account_index_map->at("account-1") + p_account_index_map->at("account-2");
     
@@ -112,11 +112,11 @@ BOOST_AUTO_TEST_CASE( test2 )
     
     double_type total_sum = content_sum + account_sum;
     
-    BOOST_CHECK_CLOSE(account_sum, 1, 1e-3);
+    BOOST_CHECK_CLOSE((double) account_sum, 1, 1e-3);
 
-    BOOST_CHECK_CLOSE(content_sum, 1, 1e-3);
+    BOOST_CHECK_CLOSE((double) content_sum, 1, 1e-3);
 
-    BOOST_CHECK_CLOSE(total_sum, 2, 1e-3);
+    BOOST_CHECK_CLOSE((double) total_sum, 2, 1e-3);
     
     
     BOOST_CHECK_EQUAL(calculator->get_total_handled_block_count(), 1);
